requirements/dev.txt
alembic==1.7.7
    # via -r requirements/./prod.in
anyio==3.5.0
    # via starlette
asgiref==3.5.1
    # via uvicorn
attrs==21.4.0
    # via -r requirements/./prod.in
autoflake==1.4
    # via -r requirements/dev.in
black==22.3.0
    # via -r requirements/dev.in
click==8.1.3
    # via
    #   black
    #   uvicorn
fastapi==0.75.2
    # via -r requirements/./prod.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
isort==5.10.1
    # via -r requirements/dev.in
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via mako
mypy==0.961
    # via -r requirements/dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
psycopg2-binary==2.9.3
    # via -r requirements/./prod.in
pydantic==1.9.0
    # via fastapi
pyflakes==2.4.0
    # via autoflake
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.36
    # via
    #   -r requirements/./prod.in
    #   alembic
starlette==0.17.1
    # via fastapi
tomli==2.0.1
    # via
    #   black
    #   mypy
typing-extensions==4.2.0
    # via
    #   mypy
    #   pydantic
uvicorn==0.17.6
    # via -r requirements/./prod.in
